+++ Readme.md
@@ -1,7 +1,7 @@
 # Личный проект «Кексобукинг»
 
 * Студент: [Ирина Дружинина](https://up.htmlacademy.ru/javascript/23/user/1086347).
-* Наставник: `Юрий Белобородов`.
+* Наставник: [Юрий Белобородов](https://htmlacademy.ru/profile/id514591).
 
 ---
