+++ index.html
@@ -231,5 +231,5 @@
   </template>
 
 </body>
-<script src="./js/main.js"></script>
+<script type="module" src="./js/main.js"></script>
 </html>
